Refactoring: darkModeButton and ToggleButton component
