+++ components/common/Loading.module.scss
@@ -26,7 +26,7 @@
 
   & > div {
     & > div {
-      color: var(--liveuta-hover-color);
+      color: var(--liveuta-header-color);
       position: relative;
       text-align: center;
