+++ components/common/Loading.tsx
@@ -9,7 +9,7 @@ const Loading = () => {
       <div>
         <div>
           <Image src={loadingImage} width={100} height={100} alt="loading_img" unoptimized priority />
-          <AiOutlineLoading size={130} color={'inherit'} />
+          <AiOutlineLoading size={130} color="inherit" />
         </div>
         <p>Loading Now!</p>
       </div>
