+++ components/common/button/ToggleButton.module.scss
@@ -1,0 +1,65 @@
+@import 'var';
+@import 'mixin';
+@import 'placeholder';
+
+.toggleButton {
+  display: inline-flex;
+  align-items: center;
+  gap: 0.5rem;
+  cursor: pointer;
+
+  input[type='checkbox'] {
+    appearance: none;
+    position: relative;
+    border: max(2px, 0.1em) solid gray;
+    border-radius: 1.25em;
+    width: 2.25em;
+    height: 1.25em;
+
+    &:before {
+      content: '';
+      position: absolute;
+      left: 0;
+      width: 1em;
+      height: 1em;
+      border-radius: 50%;
+      transform: scale(0.8);
+      background-color: gray;
+      transition: left 250ms linear;
+    }
+
+    &:checked {
+      background-color: tomato;
+      border-color: tomato;
+
+      &:before {
+        background-color: white;
+        left: 1em;
+      }
+    }
+
+    &:enabled:hover {
+      box-shadow: 0 0 0 max(4px, 0.2em) lightgray;
+    }
+
+    &:disabled {
+      border-color: lightgray;
+      opacity: 0.7;
+      cursor: not-allowed;
+
+      &:before {
+        background-color: lightgray;
+      }
+
+      & + span {
+        opacity: 0.7;
+        cursor: not-allowed;
+      }
+    }
+
+    &:focus-visible {
+      outline-offset: max(2px, 0.1em);
+      outline: max(2px, 0.1em) solid tomato;
+    }
+  }
+}
