+++ package.json
@@ -16,7 +16,6 @@
     "next": "^13.4.9",
     "pretendard": "^1.3.8",
     "react": "^18.2.0",
-    "react-dark-mode-toggle": "^0.2.0",
     "react-dom": "^18.2.0",
     "react-icons": "^4.8.0",
     "react-toastify": "^9.1.2",
